Display backends rework (#539)
* draft

* remove duplicates on src sens getters

* add nested coll support for mpl

* support nested collection for plotly

* add setters

* update set_children_styles

* update collection.remove to be search recursively

* add describe method

* adapt getBH to nested collections

* remove irrelevant todos

* add TODOS

* remove print statement

* add self to collection.describe()

* add parent/child philosophy

* fix regression

* add override_parent to Collection

* fix __add__

* add collection repr, also nested

* fix remove and make private

* testbug fix

* mini syntax sugar

* fix nested rotation

* remove unused imports

* mini color fix

* rework collection.describe()

* fix parent setter

* fix override

* fix plus, remove minus

* add describe and repr_html to all objects

* add properties

* prettify

* fix collection setter

* docstring

* docstrings

* + - behavior fixing

* replace labels with desc

* add basic tests

* include errors='raise' in remove

* only *arg object inputs for collections

* removed Collection "special" repr

* remove allows *args

* draft

* fixing and adding tests

* remove now useless radd

* fix copy parent bug

* fix tests and misc

* refactoring

* completing tests, full coverage

* pylint

* meaningful renaming

* collection and parent docstrings

* use orient concat on rotated sensors

* docstings + long forgotten properties

* make compatible with python 3.7

* changelog and pull from main

* remove describe method

* will be added in another more complete PR

* remove unused import

* coverage

* remove print statement

* pylint

* add pixel pretty describe

* allow different pix shapes with pixel_agg

* add input check

* fix input checks

* fix orientation tiling&repeating

* mini bug fix

* add tests

* fix input_checks syntax test

* fix plotly mulitidimensional pixel display

* describe, repr, tests and changelog

* draft

* remove print statement

* collection setter override parents

* remove print statement

* refactor getBHdict

* remove obsolete comment

* pylint

* core function argument consistency

* 'field' is now a positional arg only
* field and observer always in pos 1 and 2

* remove return self of reset()

* allow collection/add/remove flat list input

* take reset change back

* complete docs

* remove introduction from init

* implement field func

* fix docs

* coverage

* coverage

* extend pixel_agg testing

* fix bad tiling bug

* add test

* fix example

* fix tile_group_property

* remove print statement

* make tiling vs superposition test more general

* make tiling vs superposition test more general

* fix current_line_field call

* fix pixel_agg

* finish merge

* add pandas

* draft

* typo

* pylint

* add pandas

* update docstrings

* install .dev on circleci

* code style

* fix tests

* docstring

* move getBHdict module into level2 getBH

* refactoring

* move matplotlib folder

* remove double input checks

* refactoring

* pyvista draft implementation

* remove scalar bar

* fix remove scalar bar

* typo

* add back animation

* add matplotlib auto backend

* add colorscale caching

* remove print statement

* add cached colorscale to colormap translator

* add facecolor subdivision for matplotlib_auto

* remove bad import

* add mag arrows for generic traces

* run precommit

* fix subdivide mesh by facecolor

* renaming

* make animation generic

* handle markers in main show function

* pylint

* draft matplotlib_auto animation

* add return animation object option

* archive matplotlib old

* make extra trace generic

* remove pyvista (now generic branch)

* animation kwargs handling bug fix

* simplify structure and avoid redundancy

* fix import paths

* remove superfluous nan value in scatter mag arrow

* pylint

* lgtm

* add symb and line style translation for mpl

* fix animation with user axes

* fix output dataframe ignored

* add tests

* add changelog entry

* add example

* black

* test cov

* pylint

* fix traces grouping

* draft extra model3d compatibility

* fix tests

* modify for generic backend

* document generic backend traces

* fix bad arrow orientation with (0,0-1)

* minibug f-string fix

* docstrings minifix

* modify introduction

* fix dataframe output with pixel_agg

* remove unused test

* refactoring

* typo

* move line tiling to field function

* avoid tiling warning with vertices with diff len

* refactor field_func

* Register field functions

* automatize imports

* automatize imports

* back to explicit imports

* renaming  modules finish

* refactor param dims properties

* merge get_BH_wrap functions into one file

* pylint

* update backend docstrings

* register kind

* add matplotlib numbering

* return_fig, update_layout and layout

* fig show bug fix plotly

* Merge branch 'display-backend-rework' of https://github.com/magpylib/magpylib into refactor-getBH-02

* Merge branch 'display-backend-rework' of https://github.com/magpylib/magpylib into refactor-getBH-02

* fix bad pull from display-rework

* pylint

* draft getbdict vert feature

* fix tests

* pull from refactor_getbh

* pylint

* fix getB dict Line with "scalar" vertices

* fix mag show bug

* update compound example with generic extra trace

* remove unused files

* coverage

* fix ragged seq tiling for gebhdict line

* coverage

* coverage

* update changelog

* update changelog

* frame duration for mpl

* refactor inheritance

* move field func logic to BaseSource class

* refactor Registering classes with a metaclass

* fix subclassing sideeffect with metaclass
e.g. subclasssing Cuboid with MyCuboid, adds MyCuboid but removes Cuboid

* pylint cycling import fix

* finish pull from refactor-getBH-temp

* coverage

* small refactoring of magnet and current style
