+++ magpylib/_src/defaults/defaults_classes.py
@@ -50,7 +50,8 @@ class Display(MagicProperties):
     ----------
     backend: str, default='matplotlib'
         Defines the plotting backend to be used by default, if not explicitly set in the `display`
-        function. Can be one of `['matplotlib', 'plotly']`
+        function (e.g. 'matplotlib', 'plotly').
+        Supported backends are defined in magpylib.SUPPORTED_PLOTTING_BACKENDS
 
     colorsequence: iterable, default=
             ['#2E91E5', '#E15F99', '#1CA71C', '#FB0D0D', '#DA16FF', '#222A2A',
@@ -80,7 +81,8 @@ class Display(MagicProperties):
     @property
     def backend(self):
         """plotting backend to be used by default, if not explicitly set in the `display`
-        function. Can be one of `['matplotlib', 'plotly']`"""
+        function (e.g. 'matplotlib', 'plotly').
+        Supported backends are defined in magpylib.SUPPORTED_PLOTTING_BACKENDS"""
         return self._backend
 
     @backend.setter
