--- magpylib/_src/display/plotly/__init__.py
@@ -1,1 +1,0 @@
-"""display.plotly"""
