+++ magpylib/_src/display/sensor_mesh.py
