+++ magpylib/_src/obj_classes/class_Collection.py
@@ -120,6 +120,8 @@ def collection_tree_generator(
 class BaseCollection(BaseDisplayRepr):
     """Collection base class without BaseGeo properties"""
 
+    _draw_func = None
+
     def __init__(self, *children, override_parent=False):
 
         BaseDisplayRepr.__init__(self)
