+++ magpylib/_src/obj_classes/class_Sensor.py
@@ -1,6 +1,7 @@
 """Sensor class code
 DOCSTRINGS V4 READY
 """
+from magpylib._src.display.traces_generic import make_Sensor
 from magpylib._src.fields.field_wrap_BH import getBH_level2
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
@@ -76,6 +77,8 @@ class Sensor(BaseGeo, BaseDisplayRepr):
     """
 
     _style_class = SensorStyle
+    _autosize = True
+    _draw_func = make_Sensor
 
     def __init__(
         self,
