+++ magpylib/_src/obj_classes/class_magnet_Cylinder.py
@@ -1,10 +1,10 @@
 """Magnet Cylinder class code
 DOCSTRINGS V4 READY
 """
+from magpylib._src.display.traces_generic import make_Cylinder
 from magpylib._src.fields.field_BH_cylinder_segment import magnet_cylinder_field
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseExcitations import BaseMagnet
-from magpylib._src.style import MagnetStyle
 
 
 class Cylinder(BaseMagnet):
@@ -82,7 +82,7 @@ class Cylinder(BaseMagnet):
 
     _field_func = staticmethod(magnet_cylinder_field)
     _field_func_kwargs_ndim = {"magnetization": 2, "dimension": 2}
-    _style_class = MagnetStyle
+    _draw_func = make_Cylinder
 
     def __init__(
         self,
