+++ magpylib/_src/obj_classes/class_magnet_CylinderSegment.py
@@ -3,12 +3,12 @@ DOCSTRINGS V4 READY
 """
 import numpy as np
 
+from magpylib._src.display.traces_generic import make_CylinderSegment
 from magpylib._src.fields.field_BH_cylinder_segment import (
     magnet_cylinder_segment_field_internal,
 )
 from magpylib._src.input_checks import check_format_input_cylinder_segment
 from magpylib._src.obj_classes.class_BaseExcitations import BaseMagnet
-from magpylib._src.style import MagnetStyle
 
 
 class CylinderSegment(BaseMagnet):
@@ -91,7 +91,7 @@ class CylinderSegment(BaseMagnet):
 
     _field_func = staticmethod(magnet_cylinder_segment_field_internal)
     _field_func_kwargs_ndim = {"magnetization": 2, "dimension": 2}
-    _style_class = MagnetStyle
+    _draw_func = make_CylinderSegment
 
     def __init__(
         self,
