+++ magpylib/_src/obj_classes/class_misc_Dipole.py
@@ -1,6 +1,7 @@
 """Dipole class code
 DOCSTRINGS V4 READY
 """
+from magpylib._src.display.traces_generic import make_Dipole
 from magpylib._src.fields.field_BH_dipole import dipole_field
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseExcitations import BaseSource
@@ -80,6 +81,8 @@ class Dipole(BaseSource):
     _field_func = staticmethod(dipole_field)
     _field_func_kwargs_ndim = {"moment": 2}
     _style_class = DipoleStyle
+    _draw_func = make_Dipole
+    _autosize = True
 
     def __init__(
         self,
